moving DrawingView to its own class
