+++ app/src/main/java/course/example/ben/hobosigns/DrawingView.java
@@ -1,0 +1,140 @@
+package course.example.ben.hobosigns;
+
+import android.content.Context;
+import android.graphics.Bitmap;
+import android.graphics.Canvas;
+import android.graphics.Color;
+import android.graphics.Paint;
+import android.graphics.Path;
+import android.util.AttributeSet;
+import android.view.MotionEvent;
+import android.view.View;
+
+/**
+ * Created by enoch on 12/6/15.
+ */
+// Code adapted from http://stackoverflow.com/questions/16650419/draw-in-canvas-by-finger-android
+public class DrawingView extends View {
+    private Bitmap mBitmap;
+    private Canvas mCanvas;
+    private Path mPath;
+    public Paint mPaint;
+    Context context;
+    private Paint circlePaint;
+    private Path circlePath;
+
+    public DrawingView(Context c) {
+        super(c);
+        init(c);
+    }
+
+    public DrawingView(Context c, AttributeSet attrs) {
+        super(c, attrs);
+        init(c);
+    }
+
+    public DrawingView(Context c, AttributeSet attrs, int defStyle) {
+        super(c, attrs, defStyle);
+        init(c);
+    }
+
+    void setmBitmap(Bitmap b) {
+        mBitmap = b;
+    }
+
+    private void init(Context c) {
+        context=c;
+        mPath = new Path();
+        circlePaint = new Paint();
+        circlePath = new Path();
+        circlePaint.setAntiAlias(true);
+        circlePaint.setColor(Color.BLUE);
+        circlePaint.setStyle(Paint.Style.STROKE);
+        circlePaint.setStrokeJoin(Paint.Join.MITER);
+        circlePaint.setStrokeWidth(4f);
+    }
+
+    @Override
+    protected void onSizeChanged(int w, int h, int oldw, int oldh) {
+        super.onSizeChanged(w, h, oldw, oldh);
+
+        if (mBitmap != null) {
+            mBitmap = Bitmap.createScaledBitmap(mBitmap, w, h, true);
+        } else {
+            mBitmap = Bitmap.createBitmap(w, h, Bitmap.Config.ARGB_8888);
+        }
+        mCanvas = new Canvas(mBitmap);
+    }
+
+    @Override
+    protected void onDraw(Canvas canvas) {
+        super.onDraw(canvas);
+
+        canvas.drawBitmap(mBitmap, 0, 0, null);
+        canvas.drawPath( mPath, mPaint);
+        canvas.drawPath( circlePath,  circlePaint);
+    }
+
+    private float mX, mY;
+    private static final float TOUCH_TOLERANCE = 4;
+
+    private void touch_start(float x, float y) {
+        mPath.reset();
+        mPath.moveTo(x, y);
+        mX = x;
+        mY = y;
+    }
+
+    private void touch_move(float x, float y) {
+        float dx = Math.abs(x - mX);
+        float dy = Math.abs(y - mY);
+        if (dx >= TOUCH_TOLERANCE || dy >= TOUCH_TOLERANCE) {
+            mPath.quadTo(mX, mY, (x + mX)/2, (y + mY)/2);
+            mX = x;
+            mY = y;
+
+            circlePath.reset();
+            circlePath.addCircle(mX, mY, 30, Path.Direction.CW);
+        }
+    }
+
+    private void touch_up() {
+        mPath.lineTo(mX, mY);
+        circlePath.reset();
+        // commit the path to our offscreen
+        mCanvas.drawPath(mPath,  mPaint);
+        // kill this so we don't double draw
+        mPath.reset();
+    }
+
+    @Override
+    public boolean onTouchEvent(MotionEvent event) {
+        float x = event.getX();
+        float y = event.getY();
+
+        switch (event.getAction()) {
+            case MotionEvent.ACTION_DOWN:
+                touch_start(x, y);
+                invalidate();
+                break;
+            case MotionEvent.ACTION_MOVE:
+                touch_move(x, y);
+                invalidate();
+                break;
+            case MotionEvent.ACTION_UP:
+                touch_up();
+                invalidate();
+                break;
+        }
+        return true;
+    }
+
+    Bitmap saveImage() {
+        this.setDrawingCacheEnabled(true);
+        this.buildDrawingCache();
+        Bitmap image = Bitmap.createBitmap(this.getDrawingCache());
+        this.setDrawingCacheEnabled(false);
+
+        return image;
+    }
+}
